Move shadow rendering mode to a separate DrawerShadow class.
